remoteproc/pru: rework pru_rproc_get_id() to directly set pru id
The PRU remoteproc probe function is not returning the proper error
value upon failure of the pru_rproc_get_id() function. Rework the
pru_rproc_get_id() function to set the PRU id within the function,
and renaming the function to pru_rproc_set_id(). This simplifies the
fix in the probe function for the above problem, and also facilitates
adding back a different version of the pru_rproc_get_id() function
and export it for other PRUSS drivers/clients to retrieve a PRU id.

Fixes: b3ee6acf8eea ("remoteproc/pruss: add PRU remoteproc driver")
Signed-off-by: Suman Anna <s-anna@ti.com>