remoteproc/pruss: fix missing of_node_put() on memories node
The of_get_child_by_name() function increments the device node
reference count upon returning a successful node and requires
a matching of_node_put() to balance the reference count once
the node pointer usage is done with. This is currently missing
on the PRUSS memories child node, fix this.

Fixes: 46e809f5b457 ("remoteproc/pruss: add a platform driver for PRUSS in TI SoCs")
Signed-off-by: Suman Anna <s-anna@ti.com>