net: hsr: add directed tx support
Normally frames submitted to HSR interface are duplicated by the
HSR layer and a copy of the duplicated frame is sent out
through one slave port. This patch adds support of sending a
submitted frame through a HSR slave port specified by the io_port
field of the redundant network info submitted together with the
frame without duplicating the frame, i.e. tx as directed by the
redundant network info. In this case, the HSR tag of the out
going frame is filled in by using information in the submitted
redundant network info.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>