ARM: dts: dt-overlays: fix build failure with multiple .its files
If there are multiple .its files introduced in the system, the dependency
file generation fails miserably. The foreach statement used to generate
those, appends the last command from previous iteration to the first
command from next iteration, which obviously won't pass with the shell.
Fix by just adding a single linefeed at the end of the macro that
generates the dependencies for a single file.

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Reported-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Tested-by: Tomi Valkeinen <tomi.valkeinen@ti.com>