ARM: dts: am571x-idk: Add support for ICSS1 PRU Ethernet
AM571x-idk can have all 4 PRU Ethernet ports active simultaneously
when jumper J51 is not populated. Enable ICSS1 ethernet ports
on am571x-idk.dts.

When Jumper J51 is populated LCD on VOUT1 is possible but
ICSS1 Ethernet ports must be disabled. This configuration is
supported via am571x-idk-lcd-osd*.dts.

To allow for this we keep the bits common for both cases in
am571x-idk-common.dtsi. The ICSS1 Ethernet bits are added in
am571x-idk.dts and the LCD/touch bits are added in the resepective
am571x-idk-lcd-osd*dts.

Due to the way pins are used when all PRU Ethernet ports are
active simultaneously we need the following configuration
for GP_MUX_MODE in the ICSS.

On the board, as PR1_MII0 is brought out on pins shared with PRU2_0_GPO,
PRU2_0 GP_MUX_MODE must be something other than GP mode to avoid
contention with PR1_MII0. We select MII2 mode as it uses least
amount of GP pins.

Similarly, as PR1_MII1 is brought out on pins shared with PRU2_0_GPO
and PRU1_1_GPO, PR2_0 and PRU1_1's GP_MUX_MODE must be something other
than GP mode to avoid contention with PR1_MII1. We select MII2 mode
as it uses least amount of GP pins.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>