ARM: OMAP2+: Extend DRA7 IPU1 MMU pdata quirks to DSP MDMA MMUs
The C66-based DSPs on DRA7xx SoCs do not support a Powerdown-RET
mode, and only supports a Powerdown-Grid OFF mode which requires
a boot from reset. The HW_AUTO setting and a target power domain
state of OFF implies that the DSPs are powered off as soon as
they are idled by executing an IDLE instruction. The DSPs lose
their context as a result and will be unable to resume operations
from any wakeup event.

The DSP power domains therefore need to be restricted to ON state
for the duration a DSP processor is actively running. This is
similar to the restriction required for DRA7 IPU1 processor (albeit
because of a different reason). The IPU1 behavior is handled in
commit e978c8eba906 ("ARM: OMAP2+: use separate IOMMU pdata to fix
DRA7 IPU1 boot") which adds a .set_pwrdm_constraint ops to the OMAP
IOMMU platform data to restrict the IPU1 power domain to ON state
during the active period of the IPU1 remote processor.

Extend the IPU1 iommu pdata quirks to the DRA7 MDMA MMUs as well
to restrict the DSP power domains to ON state. The MDMA MMU module
configuration will be the first and last steps in the boot and
shutdown sequences of the DSP processors. The existing IPU1 IOMMU
pdata variable has also been renamed appropriately to reflect the
common usage between the IPU1 and the DSPs.

NOTE:
1. This patch is needed on 4.14 kernel for slightly different reasons
   due to difference in behavior between 4.14 kernel and 4.9 kernel.
   The DSP1 and DSP2 power domains are programmed for OFF target
   power state, and the power domain wakeup transition is automatic
   on any clock domain wakeup event. The DSP power domains are getting
   turned ON automatically on 4.14 kernel, but are not returned to
   OFF mode before this patch. This patch fixes this power-down issue
   by specifically programming the OFF state during the remoteproc
   shutdown. The difference in behavior is attributed to the small
   differences in commit f7b0b1e6b00a ("ARM: DRA7: hwmod data: Add
   MMU data for DSPs") on the second MMU hwmod data for each DSP.
2. This patch is also needed to preserve the DSP contexts when proper
   clock-gating is achieved during inactive periods. DSP power domains
   on these platforms should not be hitting OFF at the moment (even
   with firmware images executing an IDLE instruction) because of the
   issue described in errata i879 ("DSP MStandby requires CD_EMU in
   SW_WKUP") affecting these SoCs. The i879 errata fix in the following
   patch achieves the DSP clock-gating with HW_AUTO mode, and would
   also result in a power domain sleep transition to OFF mode without
   any software context save mechanism.
3. The functional behavior had also been inconsistent between different
   DRA74x and DRA72x SoCs and silicon revisions prior to this patch.
   DSP1 on DRA7 EVM rev.H (DRA752 ES2.0) and AM571x IDK (DRA722 ES2.0)
   boards is entering idle mode without any fix, but none of the other
   boards exhibit this behavior. They remained functional due to the
   behavioral difference in #1 on 4.14 kernel, but not on 4.9 kernel.

Signed-off-by: Suman Anna <s-anna@ti.com>