rpmsg: rpc: fix potential out-of-bounds array access
The sanity check for translation elements in rppc_xlate_buffers()
function initializes some local variables using an array index
variable prior to checking the validity of the index itself. This
can result in an out-of-bounds array access failure. Fix this by
initializing the variables after checking the array index value.

Fixes: 77e1bc79478b ("rpmsg: rpc: introduce a new rpmsg_rpc driver")
Signed-off-by: Suman Anna <s-anna@ti.com>