ARM: dts: dt-overlays: fix dependency file generation when adding new .its
The dependency file is not updated at the moment when new .its files are
added to build. Fix this by adding a dependency towards the Makefile
itself, so the dependencies get updated whenever the Makefile changes.
Also, re-fix the broken make clean functionality, this was already
fixed once, but was broken during the evolution of the dt-overlay build
mechanism.

Signed-off-by: Tero Kristo <t-kristo@ti.com>
Reported-by: Faiz Abbas <faiz_abbas@ti.com>
Tested-by: Faiz Abbas <faiz_abbas@ti.com>