arm64: dts: ti: k3-am6: Add hwspinlock node
The Main NavSS block on AM6x SoCs contains a HwSpinlock IP instance
that is similar to the IP on some OMAP SoCs. Add the DT node for
this on AM6x SoCs. The node is present within the NavSS block, and
is added as a child node under the soc0 node following the current
flat hierarchy representation of the devices.

Signed-off-by: Suman Anna <s-anna@ti.com>