ti_config_fragments/connectivity.cfg: Make MTD cmdline partition parser built-in
On K2 platforms, SPI/MTD drivers are built-in. Hence, MTD cmdline
partition parser also needs to be built-in (CONFIG_MTD_CMDLINE_PARTS=y).
Otherwise, if its a module, cmdline parser will not be available during
SPI/MTD driver probe and will fall back to OF parser(which is built-in)
which defeats whole purpose of passing mtdparts via kernel cmdline.

Signed-off-by: Vignesh R <vigneshr@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>