ARM: dts: keystone-k2g-ice: Enable network driver
This patch adds dt bindings to enable network driver on K2G
ICE boards. This consists of enabling bindings for NSS qmss,
pktdma, 2u ethss, mdio, pinmux and netcp devices as well
as DP83867 phy.

EVM hardware spec recommends to add 0.25 nsec delay in the tx
direction and 2.25 nsec delay in the rx direction for internal
delay in the clock path to be on the safer side.

The board straps RX_DV/RX_CTRL pin of on board DP83867 phy in mode 1.
The phy data manual disallows this. Add ti,dp83867-rxctrl-strap-quirk
in the phy node to allow software to enable the workaround suggested for
this incorrect strap setting. This ensures proper operation of this PHY.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>