Documentation: Update dt documentation for NSS 2u register space usage
Like NetCP 1.4/1.5, manage first region of NSS register space (Emac)
using syscon so that driver code looks similar for all variants of
netcp. So adjust the index of Switch module to index 0 and others are
calculated relative to this index. While at it also fix some typos in
the description.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>