net: netcp: ethss: k2g: use syscon for emac register space
Like other K2 devices, use syscon for emac register space
that has the subsystem version information so that the
driver can read the ss version of NSS 2u switch in a similar
way. Also fix the code not to access the sgmii registers on NSS
2u switch as it doesn't exist and rgmii h/w interface is used.
So switch to use regmap_read() to read RGMII status register
in the emac register space.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>