net: wireless: ti: wlcore: sdio: allow pm to handle sdio power
pm_runtime handles sdio power on and power off transitions.
An old workaround for trying to control the power explicitly from the
driver was in fact causing failures on suspend/resume as the mmc layer
already power the module on resume.

Valid return values for pm_runtime_get sync are 0, 1 and <0.
After resume pm_runtime_get sync returns "1" indicating the device is
Already active.
This was causing the driver to try an re-initialize an already initialized
Device and causing sdio bus failure on resume.

Remove this manual power on/off sequence as it is in-fact not needed.

Signed-off-by: Eyal Reizer <eyalr@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>