drm/tidss: add new driver for TI Keystone platforms
This patch adds a new DRM driver for Texas Instruments DSS6 IP used on
Texas Instruments Keystone K2G SoC. The DSS6 IP is a major change to the
older DSS IP versions, which are supported by the omapdrm driver, and
while on higher level the DSS6 resembles the older DSS versions, the
registers and the internal pipelines differ a lot.

DSS6 IP on K2G is a "ultra-light" version, and has only a single plane
and a single output. The driver will also support future DSS versions,
which support multiple planes and outputs, so the driver already has
support for those.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>

To some changes were needed to backport this for ti-linux-4.14.y.

- DRM on v4.14 Linux needs still tidss_lastclose() and
  tidss_output_poll_changed() callbacks, add them

- drm_dev_put() is drm_dev_unref() on v4.14 DRM

- v4.14 DRM does not have drm_atomic_helper_check_plane_state(), use
  drm_plane_helper_check_state() instead

Signed-off-by: Jyri Sarha <jsarha@ti.com>