Revert "xhci: Reset Renesas uPD72020x USB controller for 32-bit DMA issue"
This reverts commit 8466489ef5ba48272ba4fa4ea9f8f403306de4c7.

There is a bug in Renesas PCI-USB cards where the controller retains
DMA programming across XHCI reset. If the device has been using 64 bit
DMA addresses and then switches to 32 bit addresses, the top 32 bits
are corrupted. The above patch implements a workaround by doing a PCI
function reset for all devices. However, the PCI reset causes kernel
hangs on some boards. Due to this, it was decided that the workaround
should be restricted to devices which have an iommu.
Therefore, revert the above patch.

Link to discussion:
https://lkml.kernel.org/r/6ed573a9-418f-b1da-4ae6-0ed4f41fb66e@arm.com

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>