remoteproc/pruss: select MFD_SYSCON for PRUSS
The PRUSS platform driver uses syscon nodes for couple of sub-modules
like IEP, MII_RT etc and relies on the syscon driver being enabled to
probe properly. So, make sure the MFD_SYSCON Kconfig option is enabled
whenever the PRUSS remoteproc driver is enabled.

The MFD_SYSCON is a boolean option, and the select dependency option
is chosen following the majority usage style within the kernel, as well
as to align with similar usage in couple of other remoteproc drivers.

Reported-by: Roger Quadros <rogerq@ti.com>
Fixes: ac0852d14216 ("remoteproc/pruss: parse various syscon nodes and store their regmaps")
Signed-off-by: Suman Anna <s-anna@ti.com>