dt-bindings: mailbox: ti,message-manager: Fix interrupt name error
commit 3b73ad33873e3f9ed2fe06374cef03ec463e0285 upstream.

Message Manager's mailbox interrupts are queue based and not proxy
specific. The interrupt names are wrong in the binding, however
correctly reflected in the example provided. Remove the relation
to proxy ID in the documentation of binding. Existing device tree
descriptions follow the correct conventions already and documentation
update has been missed.

Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Rob Herring <robh@kernel.org>