arm64: dts: ti: k3-am6: Add mailbox cluster nodes
The AM6x Main NavSS block contains a Mailbox IP instance with
multiple clusters. Each cluster is equivalent to an Mailbox IP
instance on OMAP platforms.

Add all the Mailbox clusters as their own nodes under the soc
node instead of creating an almost empty parent node for the
new K3 mailbox IP and the clusters as its child nodes. All these
nodes are marked as disabled, and they need to be enabled along
with the appropriate child nodes on a need basis.

NOTE:
The NavSS only has a limited number of interrupts, so all the
interrupts generated by a Mailbox IP are not added by default.
Only the needed interrupts that are targeted towards the A53
GIC will need to be be added later on when some sub-mailbox
child nodes are added.

Signed-off-by: Suman Anna <s-anna@ti.com>