arm64: dts: ti: k3-am654-base-board: Add property to handle eMMC card detect quirk
The eMMC card detect and write protect lines are not connected to
the card on the board. However, the host controller requires the
card detect line to be low to work reliably.

Workaound this by using the card detect test field to force the
card detect to 1.

This enables us to remove the unused pinmux configurations.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>