drm/omap: fix plane's set/get_property
omap_plane_atomic_set_property() and omap_plane_atomic_get_property()
use plane->state as the state to get/set the property. However, this is
the "old" state, not the new one passed to the function. With atomic
modesetting this doesn't seem to cause problems, but with legacy
modesetting the newly set property value gets discarded.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>