dmaengine: ti: k3-udma: Fix race of packet completion in while irqs are off
To handle the following scenario in client (UART for example) drivers:
queue RX transfer.
Before the packet is completed we get timeout from the peripheral
In timeout interrupt handler:
 pause the current RX and get statistics
 terminate the current RX
 queue new RX transfer

The first RX descriptor is moved to terminated_desc and the new RX is now
the desc.
After the interrupts are enabled again, we get the first RX packet back
(partial packet) followed by a teardown complete message. This is _not_
for the current desc, bot for the terminated one...

With a simple logic we can figure this out and make sure that we handle
this use case correctly.

Reported-by: Vignesh R <vigneshr@ti.com>
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>