dmaengine: ti: k3-udma: Set rx_error_handling in rflow config
We don't want UDMA to drop (return packets to free descriptor ring) in case
of descriptor starvation. Instead we want UDMA to wait until it has
sufficient amount of descriptor to complete the packet.

This can happen when a streaming type of peripheral is sending data and the
driver uses one active rx request at the time. If the delay between a
transfer completion and the start of the next rx transfer is 'long' the
channel will go through descriptor starvation.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>