dmaengine: ti: k3-udma: Resume the paused rchan when stopping it
If the channel is paused at the time when we stop it, it is not going to
move to resumed state as the rx teardown is initiated on the source, not
on the destination.

This means that after the channel teardown completion the UDMA rchan will
remain in paused state.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>