dmaengine: ti: k3-udma: Use different counters for packets and TRs
Audio is using cyclic TR mode and we used the sg_idx to track which TR just
given the interrupt (period irq). But the same sg_idx is used by the packet
tracking logic as well.

When the cyclic TR is stopped we might (or might not) get back the TR
descriptor from UDMA and if the sg_idx was not 0, we might be accessing to
invalid memory causing crash.

Use different counters for packets (within a transfer - in reality we only
have 1 packet for any given transfer type) and for the TR tracking.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>