dmaengine: ti: k3-udma: Fix channel teardown handling
In terminate_all we set the given channel to start a teardown. New
transfers can not be started before the teardown is completed and we must
not reset the RT_CTL registers before the teardown is completed.

To ensure that we obey this rule:
Wait for a completion in udma_synchronize() before resetting the RT_CTL
registers.

In issue_pending() mkae sure that we are not pushing new descriptor to the
ring while the channel is in teardown (we initiated the teardown and the
channel is still enabled).

When we receive the teardown complete message make sure that we pick up
the next pending (if any) descriptor.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>