remoteproc/pru: fix ti,pru-interrupt-map range checks for K3 SoCs
The ICSSG instances on AM65x SoCs have more System Events, Interrupt
Channels and Host Interrupts compared to the previous generation PRUSS
instances on other SoCs. The logic in pru_rproc_get() function to parse
and program the interrupts from DT is still checking against the values
relevant to previous generation PRUSS instances and therefore would
result in invalid failures when any of the new events are used on AM65x
platforms.

Fix this logic properly for K3 SoCs. The logic was simply designed around
storing a flag that is set only on K3 SoCs, and can be reworked to use
match data to scale beyond current SoCs if needed.

Fixes: 529c7767b4f3 ("remoteproc/pru: add support for various PRU cores on K3 AM6x SoCs")
Signed-off-by: Suman Anna <s-anna@ti.com>