serial: 8250_omap; Support EFR2 register
AM654 SoC has a newer version of 8250 OMAP UART IP that as Enhanced
Feature Register 2 (EFR2). This register provides a way to configure UART
RX timeout to fire when there is no activity on RX line, irrespective
whether are not there is data in RX FIFO. Using this feature its
possible to queue RX DMA for length > RX FIFO size and still reliably
detect timeout condition.

Add a quirk in the driver to make use of this feature.

Signed-off-by: Vignesh R <vigneshr@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>