arm64: dts: ti: Describe the SERDES clock inputs, outputs and mux in dt
Keep the default configuration (k3-am6.dtsi) such that PLL mux of
SERDES0 uses the the left input connected to SoC clock and PLL mux
of SERDES1 uses the right input connected to SoC clock.

In the DT overlay file for 2lane PCIe SERDES personality card,
change the configuration such that the PLL mux of SERDES1 uses the
right output of SERDES0 and the right output of SERDES0 is connected
to the left input of SERDES0.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>