iommu/omap: Fix cache flushes on L2 page table entries
The flush_iopte_range() function is used for flushing the programmed
L1/L2 page table entries from the cache. This function takes in a base
dma address, an offset from that base, and number of entries to be used
for computing the address and size to be flushed.

The base address used for DMA operations on the second-level table did
incorrectly include the offset for the table entry. The L2 page table
entry offset was then added again to the base address resulting in
flushing the cache for the wrong memory locations. This resulted in
occasional MMU faults getting triggered (even though the entries
themselves are programmed correctly) when the corresponding address
is accessed on the remote processor side when the hardware page table
walk is performed and fetching an incorrect translation.

Operations on the L1 table are not affected.

Fix this by changing base address to point to the beginning of the L2
page directory address.

Fixes: a313fc4938b8 ("iommu/omap: Use DMA-API for performing cache flushes")
Acked-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Ralf Goebel <ralf.goebel@imago-technologies.com>
[s-anna@ti.com: backport linux-omap patchwork id '10552411', revise description]
Signed-off-by: Suman Anna <s-anna@ti.com>