Revert "sched/core: Require cpu_active() in select_task_rq(), for user tasks"
This reverts commit 0d5e04e239ad5b18c4099ef942843bf510af1122.

This is needed for poweroff to work on dual core platforms Like
DRA7 & AM572x-IDK. This patch is preventing the i2c irq bottom
half from executing on dual core SoCs. Reverting this for now
is a work aorund until an upstream solution is found where
irq-less i2c transfers are supported for irq_disabled context
i2c writes, which is the case with PMICs on the above mentioned
platforms.

Signed-off-by: Keerthy <j-keerthy@ti.com>