arm64: dts: ti: Fix wrong display enable gpio and touch reset gpio
The gpio pins on pca9555 are numbered from 0 to 17, skiping 8 and
9. Then our schematics number the pins from 1 to 17, skiping only
9. But the pca9555 driver numbers them just linearly from 0 to 15. So
the gpio pin that is numered 10 in schematics, is actually number 8
when the pca9555 gpio phandle is concerned.

A similar fix is applied for touch screen reset gpio. The gpio pin 11
in the schematics translates to gpio 9 in pca9555 gpio phandle.

Signed-off-by: Jyri Sarha <jsarha@ti.com>