net: ti: icssg-prueth: Fix lock-up on TX teardown
In tx_irq handler() we don't do anything if !netif_running(),
which is the case when we're in emac_ndo_stop(). At emac_ndo_stop()
on TX teardown we get teardown completion IRQ and do nothing about it
and so just keep getting called back into the interrupt handler
indefinitely thus blocking the system and causing a RCU stall.

So there are 2 options to fix this.
1) disable_irq() before doing a teardown.
OR
2) process teardown completion IRQs even if !netif_running().

We use the 2nd option and tell k3_nav_udmax_tdown_tx_chn()
that we will process the teardown completion asynchronously.

Reviewed-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>