irqchip: ti-sci-inta: Use vint as hwirq
When creating a new IRQ from INTA irq domain, the hwirq is derived
from first requested source id and source index. Further more events
can be attached to the same irq. And the events can be released in
any order. So the event that started the irq can be released
without disposing the irq. So when creating a new irq with
this released event, there is a hit of hwirq causing irq_create
to crash. So use vint as hwirq instead of deriving it from
source id and index.

Reported-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>