dmaengine: ti: k3-udma: Fix error when channel request fails
When channel request fails the DMAengine core will call the synchronize
callback of the driver.

We should only wait for completion if we have initiated the teardown.

If the channel request fails the resources are cleared up and no tchan or
rchan is associated anymore. Check for valid tchan/rchan pointers in the
reset_ring function to handle this case cleanly.

A released PSI-L link shound not be released twice.

Lastly initialize the teardown related flag and completion early enough
since at this point the channel is not undergoing a teardown.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>