remoteproc/k3-r5: add a remoteproc driver for MCU (dual-R5F) subsystem
The AM65x SoCs have a single Microcontroller Unit (MCU) subsystem that
is comprised of dual-core Arm Cortex-R5F processor cores. This MCU R5
cluster is present within the MCU domain, and can be configured at boot
time to be either run in a LockStep mode or in an Asymmetric Multi
Processing (AMP) fashion in Split-mode. This subsystem has 64 KB each
Tightly-Coupled Memory (TCM) internal memories for each core split
between two banks - TCMA and TCMB (further interleaved into two banks).
The subsystem does not have an MMU, but a Region Address Translater
(RAT) module that is accessible only from the R5s for providing
translations between 32-bit CPU addresses into larger system bus
addresses.

Add a remoteproc driver to support this subsystem to be able to load
and boot the R5 cores primarily in LockStep mode. Error Recovery and
Power Management features are not currently supported. Loading support
includes the internal TCMs and DDR. The code includes preliminary
support for Split mode, with the full features/functionality to be
added in a future patch. The subsystem is represented as a single
remoteproc in LockStep mode, and as two remoteprocs in Split mode. The
driver uses various TI-SCI interfaces to talk to the System Controller
(DMSC) for managing configuration, power and reset management of these
cores. IPC between the A53 cores and the R5 cores is supported through
the virtio rpmsg stack using shared memory and OMAP Mailboxes.

The driver expects the following firmware names to be used for these
cores:
    am65x-mcu-r5f0-fw (LockStep mode and for Core0 in Split mode)
    am65x-mcu-r5f1-fw (Core1 in Split mode)

Signed-off-by: Suman Anna <s-anna@ti.com>