dmaengine: ti: k3-udma: Fix missing tdcm caused kernel oops
It was assumed that the tdcm will always going to be received as per the
documentation, but under unknown circumstances this is not the case.

We receive back the terminated descriptor, but the expected tdcm is not
coming as the teardown fails to complete.

In this case the cleanup code would try to free up the terminated
descriptor again, which will trigger kernel oops.

To handle this, check if the current descriptor to be freed is the
terminated one and reset the terminated_desc pointer to NULL.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>