drm/tidss: separate functions for common area
tidss supports devices which contain multiple common
areas. The behavior or the common areas and their
functionalities differ from one device to another.

In J721E, all of the common areas can be used
for handling interrupts and the first common area can
be used for global configuration of DSS.

The functionality of the COMMON region has been
separated into configuration and interrupt handling.

For now, the "common" registers are mapped into
base_common_cfg and the address is copied to
base_common_intr, which means the same area is used
for interrupt handling and global configuration
across all devices.

Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>