drm/tidss: static partition of vp and planes
DSS7 has support for partitioning the video ports
and video pipes among the several compute entities
and they can be independently driven from different
cores.

This patch adds support for statically partitioning
the video ports and pipelines using device-tree nodes.

If there is a DT node that specifies which pipes and
video ports to use, then the information is
used to populate the CRTCs and planes accordingly

Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>